backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
fastapi-users[sqlalchemy]==13.0.0
httpx-oauth==0.15.0
slowapi==0.1.9
resend==2.3.0
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
alembic==1.13.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2
aiosqlite==0.20.0
scrapy==2.11.0
boto3==1.34.0
pypdf==4.0.0
pytesseract==0.3.13
Pillow==12.1.1
psycopg[binary]==3.2.13
pdf2image==1.17.0
legacy-cgi==2.6.4
pgvector==0.3.6
anthropic==0.40.0
openai>=1.0.0
google-generativeai>=0.8.0
